# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AstroMystic API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface HoroscopeDomain {
  label: string;
  icon?: string;
  score: number;
}

export interface Horoscope {
  sign: string;
  date: string;
  message: string;
  emoji?: string;
  domains: HoroscopeDomain[];
}

export interface CompatibilityDomain {
  label: string;
  score: number;
}

export interface Compatibility {
  sign1: string;
  sign2: string;
  overallScore: number;
  message: string;
  domains: CompatibilityDomain[];
}

export interface CompatibilityUsage {
  freeTestUsed: boolean;
  isPremium: boolean;
  /** @nullable */
  testsRemaining?: number | null;
}

export interface PaymentRequired {
  error: string;
  requiresPremium: boolean;
}

export type ConsultantStatus = typeof ConsultantStatus[keyof typeof ConsultantStatus];


export const ConsultantStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Consultant {
  id: number;
  userId?: string;
  name: string;
  speciality: string;
  bio: string;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  zodiacSign?: string | null;
  /** @nullable */
  rating?: number | null;
  reviewCount?: number;
  /** @nullable */
  pricePerSession?: number | null;
  status: ConsultantStatus;
  /** @nullable */
  experience?: string | null;
  tags?: string[];
}

export interface ConsultantApplication {
  name: string;
  speciality: string;
  bio: string;
  zodiacSign?: string;
  experience?: string;
  pricePerSession?: number;
  tags?: string[];
}

export interface UserProfile {
  userId: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  zodiacSign?: string | null;
  /** @nullable */
  birthDate?: string | null;
  isPremium: boolean;
  freeTestUsed?: boolean;
  createdAt?: string;
}

export interface UserProfileUpdate {
  zodiacSign?: string;
  birthDate?: string;
  name?: string;
}

export type CheckoutRequestPlan = typeof CheckoutRequestPlan[keyof typeof CheckoutRequestPlan];


export const CheckoutRequestPlan = {
  monthly: 'monthly',
  annual: 'annual',
} as const;

export interface CheckoutRequest {
  plan: CheckoutRequestPlan;
}

export interface CheckoutSession {
  url: string;
}

export interface SubscriptionStatus {
  isPremium: boolean;
  freeTestUsed: boolean;
  /** @nullable */
  subscriptionId?: string | null;
  /** @nullable */
  currentPeriodEnd?: string | null;
}

export interface ErrorResponse {
  error: string;
}

export type GetCompatibilityParams = {
sign1: string;
sign2: string;
};