# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AstroMystic API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /consultants:
    get:
      operationId: listConsultants
      tags: [consultants]
      summary: List all approved consultants
      responses:
        "200":
          description: List of consultants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Consultant"
    post:
      operationId: applyAsConsultant
      tags: [consultants]
      summary: Apply to become a consultant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsultantApplication"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultant"
        "409":
          description: Already applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /consultants/my-profile:
    get:
      operationId: getMyConsultantProfile
      tags: [consultants]
      summary: Get the current user's consultant profile
      responses:
        "200":
          description: Consultant profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultant"
        "404":
          description: Not a consultant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Consultant:
      type: object
      required: [id, name, speciality, bio, status]
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        speciality:
          type: string
        bio:
          type: string
        avatar:
          type: ["string", "null"]
        zodiacSign:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        pricePerSession:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        experience:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
    ConsultantApplication:
      type: object
      required: [name, speciality, bio]
      properties:
        name:
          type: string
        speciality:
          type: string
        bio:
          type: string
        zodiacSign:
          type: string
        experience:
          type: string
        pricePerSession:
          type: integer
        tags:
          type: array
          items:
            type: string
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string