# lib/api-zod/src/generated/types/consultant.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AstroMystic API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ConsultantStatus } from './consultantStatus';

export interface Consultant {
  id: number;
  userId?: string;
  name: string;
  speciality: string;
  bio: string;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  zodiacSign?: string | null;
  /** @nullable */
  rating?: number | null;
  reviewCount?: number;
  /** @nullable */
  pricePerSession?: number | null;
  status: ConsultantStatus;
  /** @nullable */
  experience?: string | null;
  tags?: string[];
}