# lib/api-zod/src/generated/types/consultantApplication.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AstroMystic API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ConsultantApplication {
  name: string;
  speciality: string;
  bio: string;
  zodiacSign?: string;
  experience?: string;
  pricePerSession?: number;
  tags?: string[];
}